meson: add back warning when default value for external properties are used
This is adding back a warning that was removed in the upgrade to Meson
0.58 (cb2a7dcc) as source evolution meant the patch didn't apply.

Meson scripts which want to identify runtime behaviour often use
external properties (meson.get_external_property(), set via the cross
and native files) to define default behaviour if they can't run code at
configure time.

These defaults may or may not be correct, so emit a warning that this is
happening. The recipe should then provide a cross file fragment which
sets the value appropriately.

In the long term I hope to be able to remove this patch as Meson evolves
cc.run() to support fallback values directly, and Meson will log the use
of defaults itself.

Signed-off-by: Ross Burton <ross.burton@arm.com>