update_mandb: deb fails due to missing man cache
This only occurs in debian package builds when populating the sdk
and is a work around that seems to work. Eventually we should look
at why this is failing (I have ideas, it's somewhere in
lib/oe/package_management/deb/sdk.py), but for now, do this so we can
fix the core issue with nativesdk-intercepts.

Signed-off-by: Eilís 'pidge' Ní Fhlannagáin <pidge@baylibre.com>