oeqa/runtime/parselogs: load ignores from disk
Instead of hardcoding the list of ignored errors/warnings in the test
itself, read them plain text files on disk.

This uses importlib.resources() to try to open files called
oeqa.runtime.cases.parselogs-ignores-[candidate].txt, where the candidate
will be:

- "common"
- The TARGET_ARCH
- Each of the MACHINEOVERRDES

This allows the common and tune-specific ignores to be retained in
oe-core, and machine-specific ignores added to the layer where the
machine is defined.

To support Python 3.8, this needs a __init__.py to be created in the
cases directory so that older importlib considers it. This is not needed
with 3.9 onwards and can be removed when we required 3.9+.

[ YOCTO #14604 ]

Signed-off-by: Ross Burton <ross.burton@arm.com>