oeqa/runtime/parselogs: migrate ignores
Move the ignores from a huge dict in the parselogs.py module to .txt
files. This is just the common, tune, and qemu machine ignores; the
machine ignores that are not in oe-core will be added to the relevant
layer.

The list of ignores has not been reviewed in any meaningful way, this
should be one soon as I suspect a large number of these are redundant.

Signed-off-by: Ross Burton <ross.burton@arm.com>