fetch2/wget.py: work quietly when checking upstream versions
'devtool check-upgrade-status' floods the console with all of
the wget invocations. This is due to the behaviour of the standard
logger/progresshandler used by the wget fetcher which prints to stdout
everything wget executable outputs.

Let's use the logger only when downloading, and not when checking upstream versions.

Also, there is no need to pass the quiet option to runfetchcmd()
as runfetchcmd() has been ignoring that since 2012:

https://git.yoctoproject.org/poky/commit/?id=2b918b623bf8add454c3726dd4ee189230e6fcbc

The 'quiet' argument has since been superseded by the 'log' argument, and ideally
the quiet option should be removed altogether from the runfetchcmd() arguments,
but that would break the fetcher API.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>