oeqa/runtime/parselogs: ignore jitterentropy health failure
Sometimes the jitterentropy RNG fails to initialise on boot.  My hunch
is that this is due to the hardware timer not being high resolution
enough combined with running inside a virtualised machine on a loaded
host, and whilst the message looks bad it basically says "driver not
suitable for hardware".

I've contacted the author of the jitterentropy code to confirm or deny
this hunch, but until then we can easily ignore the failure.

(From OE-Core rev: ee205911d517c10532d0bd41e30b486c637c1bb1)

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>