package: do_package depends on binutils for objdump
do_package uses objdump to read information from binaries, but this is
removed from the default dependencies with INHIBIT_DEFAULT_DEPS.

It's possible to ship prebuilt ELF files and need objdump/strip to
package correctly, which will fail (which is partly why bin_package does
not inhibit dependencies). Also the shared library dependency logic
silently fails if objdump can't be found, which isn't ideal.

Resolve both of these by adding binutils as a dependency of the package
task, as it provides objcopy, objdump, and strip.

Signed-off-by: Ross Burton <ross.burton@arm.com>