python3-hypothesis: upgrade 6.88.3 -> 6.89.0
https://hypothesis.readthedocs.io/en/latest/changes.html#v6-89-0
https://hypothesis.readthedocs.io/en/latest/changes.html#v6-88-4

6.89.0 - 2023-11-16
This release teaches from_type() to handle constraints implied by the
annotated-types package - as used by e.g. Pydantic. This is usually
efficient, but falls back to filtering in a few remaining cases.

Thanks to Viicos for pull request #3780!

6.88.4 - 2023-11-13
This patch adds a warning when @st.composite wraps a function annotated
as returning a SearchStrategy, since this is usually an error (issue #3786).
The function should return a value, and the decorator will convert it to a
function which returns a strategy.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>