classes-global/sstate: Skip sstate mirror files that are not known to the server
sstate mirrors can be very helpful, but in normal operation clients will
make a bunch of queries to the server for hash files that don't exist.
This can put a lot of undue stress on the server, since negative lookups
can often be expensive (especially when using a CDN).

To help mitigate this, the hash equivalence server can be used to cull
the lookups that are unlikely to succeed. This is done by first querying
if the server is aware of a given unihash (which the server is optimized
to do quickly), and only if the server thinks a given unihash exists will
an attempt be made to pull it from an sstate mirror.

Note that this _only_ affect sstate mirrors; local sstate caches are
assumed to be fast for negative lookups, but we may be able to extend
this to cover the local cache also in the future if desired.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>