hashserv: sqlite: Ensure sync propagates to database connections
When the sqlite database backend was restructured, the code to make the
databases run in WAL mode and to control if sync() is called was
accidentally dropped. This caused terrible database performance to the
point that server timeouts were occurring causing really slow builds.

Fix this by properly enabling WAL mode and setting the synchronous flag
as requested

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>