ldconfig-native: Add usrmerge support
If DISTRO_FEATURE usrmerge is enabled, SLIBDIR is just a symlink to LIBDIR.
ldconfig-native uses hardcoded paths for SLIBDIR and LIBDIR that are fine for
the not usrmerge case. But if usrmerge is enabled ldconfig-native generates a
ld.so.cache for the target, which is done in rootfs.py, that looks like this:
236 libs found in cache `/etc/ld.so.cache'
       libzstd.so.1 (libc6) => /lib/libzstd.so.1
       libz.so.1 (libc6) => /lib/libz.so.1
       libxtables.so.12 (libc6) => /lib/libxtables.so.12
       libxslt.so.1 (libc6) => /lib/libxslt.so.1
       libxml2.so.2 (libc6) => /lib/libxml2.so.2

But it should look like this, which it does in case of not usrmerge:
236 libs found in cache `/etc/ld.so.cache'
        libzstd.so.1 (libc6) => /usr/lib/libzstd.so.1
        libz.so.1 (libc6) => /usr/lib/libz.so.1
        libxtables.so.12 (libc6) => /usr/lib/libxtables.so.12
        libxslt.so.1 (libc6) => /usr/lib/libxslt.so.1
        libxml2.so.2 (libc6) => /usr/lib/libxml2.so.2

As this is part of the generated image(target) this breaks for example the
Qt QLibraryInfo paths if Qt is configured 'relocatable' and leads to the issue
that the QtWebEngine can find the QtWebEngineProcess,
Qt WebEngine resources,...
In the case of Qt this issue appears because Qt tries to determine the library
locations using dladdr and uses this as base to access QtWebEngineProcess or
Qt Webengine resources.

Therefore don't let ldconfig-native parse SLIBDIR if it is a symlink.

Signed-off-by: Johannes Pointner <johannes.pointner@br-automation.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>