meta/conf/templates/default/conf-description.txt: add a template description
Existing conf-notes.txt is not adequate in situations where the user
needs to select a config template from several, or get an overview of them and
is interested only in a brief explanation of what the template is for,
but not extended multi-paragraph instructions for how to use it.

Such descriptions can be provided via conf-description.txt and
this patch adds such a description for the oe-core default template.

Updates to the tools and a tool for selecting templates which
will make specific use of the description will follow.

(From OE-Core rev: a2d56474586b011d238c3f7e371dc9a1721de9bb)

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>