image.bbclass/rootfs: set and unpack package-database
set the package-database of a "lower image" to unpack and build upon when
installing packages for the current image. This way a lean image will be
created, which only holds the packages that are not already present in the lower
image, that then could be used with overlayfs or systemd-sysext to extend the
"lower image" on demand; for development purposes on an RO lower image for
example.

(From OE-Core rev: 3fc3816171e931323ec3f82a5ce6d0b4aaad9cd7)

Signed-off-by: Johannes Schneider <johannes.schneider@leica-geosystems.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>