hashserv: Report supported API from server
Enables support for the hash server to report which APIs it implements
to clients in the server headers. This can be used in the future to
allow newer clients to negotiate correctly with older servers if newer
APIs are added.

NOTE: This is a breaking change on the client side that mandates server
updates in order for clients to work, but should hopefully prevent
needing to do this in the future

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>